requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.5.0

# LangChain
langchain>=0.0.352
langchain-community>=0.0.14
langchain-deepseek>=0.0.1

# Vector Database
pymilvus>=2.3.1
faiss-cpu>=1.7.4

# Embeddings
sentence-transformers>=2.2.2

# Document Processing
unstructured>=0.11.0
pdf2image>=1.16.3
pytesseract>=0.3.10
pillow>=10.1.0
beautifulsoup4>=4.12.2
requests>=2.31.0
chardet>=5.2.0
docx2txt>=0.8

# Data Processing
numpy>=1.26.0
pandas>=2.1.2

# Logging and Monitoring
loguru>=0.7.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Development
black>=23.10.1
isort>=5.12.0
flake8>=6.1.0
